Support for two-leg removed; Use simple-ipam
